Added normalization
Added a normalization option to hamming-distance and
levenshtein-distance. If supplied and not false, the distance value is
normalized to a value in the range between 0 and 1 (both ends
inclusive).

Fix: the jaro and jaro-winkler distances were broken. The current
versions required, that the longer sequence was always passed as the
second argument. This has been fixed: now, the implementation no longer
cares, whether one sequence is longer than the other.